Bad truncate messge structure   Structure sp_truncate_arg_t has a field last_seg of size 32 bits while   structure sp_truncate_arg_no_bins_t has this field of size 16 bits.   The filed is actualy access thanks to sp_truncate_arg_no_bins_t by the   STORCLI as well as the STORIO.   Structure sp_truncate_arg_t has been fixed.

+++ rozofs/rpc/sproto.h
@@ -220,7 +220,7 @@ struct sp_truncate_arg_t {
 	uint32_t dist_set[ROZOFS_SAFE_MAX_RPC];
 	sp_uuid_t fid;
 	uint8_t proj_id;
-	uint32_t last_seg;
+	uint16_t last_seg;
 	uint64_t last_timestamp;
 	uint64_t bid;
 	struct {
